Allow multiple package specifiers at the same time
This allows e.g:
    $ hashin packageA==1.2.3 packageB==2.3.4 packageC

This change means that 'requirements_file' and 'algorithm' can
no longer be positional arguments, so -r/--requirements-file and
-a/--algorithm were added for them respectively. Obviously this
breaks backwards compat.

This is also the simplest way to implement it, but not necessarily
the most efficient. For example, this will open/close the requirements
file each time, rather than leaving it open for the duration.