Add an optional --index-url argument similar to that in pip (#107)
Adds an optional `--index-url` argument so that hashin can be used with an alternate package index.

Pip supports this behavior with the same argument. Hashin was hardcoded to point at pypi.org, so if you happened to be working with an alternate index, hashin couldn't help you.

Pip also accepts an `--extra-index-url` argument, but I'm guessing that hashin won't benefit much from supporting the same. I expect the main use case of `--index-url` to be a package or package version that only exists on the alternate index.