Fixes #116
This patch changes hashin such that,

If the package specified has underscores in its name, but the new package
has hyphens, the following occurs

* If there is a new version, hashin updates the package in requirements to
  be the hyphenated name
* If there is no new version, the underscored name remains in place

Prior to this patch, the 1st case about would not occur. Instead, hashin
would append the new version in its hyphenated form to the requirements file
and leave the old version in its underscored form in the requirements file.